make geopandas and other heavy deps optional, reorganization
