+++ sewergraph/__init__.py
@@ -9,7 +9,7 @@ from .core import *
 from .helpers import *
 from .area_calcs import *
 from .resolve_data import *
-
+from sewergraph.save_load import graph_from_shp, gdf_from_graph, graph_from_gdf
 
 VERSION_INFO = (0, 1, 3)
 __version__ = '.'.join(map(str, VERSION_INFO))
