+++ sewergraph/core.py
@@ -1,83 +1,6 @@
 import networkx as nx
-import pandas as pd
-import geopandas as gp
-from .helpers import pairwise, generate_facility_id, get_node_values
+from .helpers import pairwise, get_node_values
 from .hhcalculations import philly_storm_intensity
-
-
-def graph_from_shp(pth=r'test_processed_01', idcol='facilityid', crs={'init': 'epsg:4326'}):
-    # import well-known-text load func
-    from shapely import wkt
-
-    G = nx.read_shp(pth, )
-    G.graph['crs'] = crs
-    G = nx.convert_node_labels_to_integers(G, label_attribute='coords')
-
-    for u, v, d in G.edges(data=True):
-
-        # create a shapely line geometry object
-        d['geometry'] = wkt.loads(d['Wkt'])
-        d['length'] = d['geometry'].length
-
-        # get rid of other geom formats
-        del d['Wkb'], d['Wkt'], d['Json']
-
-        # generate a uniq id if necessary
-        if idcol not in d:
-            d[idcol] = generate_facility_id()
-
-    return G
-
-
-def gdf_from_graph(G, return_type='edges'):
-    '''create a GeoDataFrame from a sewergraph, G.'''
-    if return_type == 'edges':
-        df = nx.to_pandas_edgelist(G)
-        return gp.GeoDataFrame(df, crs=G.graph['crs'])
-
-    elif return_type == 'nodes':
-        node_dict = {n: d for n, d in G.nodes(data=True)}
-        df = pd.DataFrame(node_dict).T
-        return gp.GeoDataFrame(df, crs=G.graph['crs'])
-    else:
-        print('incorrect return type. should be "edges" or "nodes".')
-
-
-def graph_from_gdf(gdf):
-    '''create a sewergraph, G, from a GeoDataFrame'''
-    G = nx.from_pandas_edgelist(gdf, create_using=nx.DiGraph(), edge_attr=True)
-    G.graph['crs'] = gdf.crs
-    return G
-
-
-def transform_projection(G, to_crs='epsg:4326'):
-    '''
-    change the coordinate projection of Shapely "geometry" attributes in edges
-    and nodes in the graph
-    '''
-    from shapely.ops import transform
-    from functools import partial
-    try:
-        import pyproj
-    except ImportError:
-        raise ImportError('pyproj module needed. get this package here: ',
-                          'https://pypi.python.org/pypi/pyproj')
-
-    # set up the projection parameters
-    st_plane = pyproj.Proj(G.graph['crs'], preserve_units=True)
-    wgs = pyproj.Proj(init=to_crs)  # google maps, etc
-    project = partial(pyproj.transform, st_plane, wgs)
-
-    # apply transform to edge and node geometry attributes
-    for u, v, geometry in G.edges(data='geometry'):
-        if geometry:
-            G[u][v]['geometry'] = transform(project, geometry)
-
-    for n, geometry in G.nodes(data='geometry'):
-        if geometry:
-            G.node[n]['geometry'] = transform(project, geometry)
-
-    G.graph['crs'] = to_crs
 
 
 def hydrologic_calcs_on_sewers(G, nbunch=None, return_period=0):
